+++ src/components/layout/Layout.css
@@ -10,6 +10,20 @@ footer {
     background: rgb(39, 26, 6);
 }
 
+header a {
+    color: white;
+    font-size: 1.3rem;
+    text-decoration: none;
+    transition: 0.3s;
+}
+
+.active {
+    padding: 5px;
+    color: black;
+    font-weight: bold;
+    background-color: orange;
+}
+
 main {
     display: flex;
     justify-content: center;
